leetie: Add 396. Rotate Function [Medium] (java)

+++ solutions/Medium/0396-rotate-function/solution.java
@@ -4,7 +4,7 @@
 // Tags     : Array, Math, Dynamic Programming
 // Link     : https://leetcode.com/problems/rotate-function/
 // Runtime  : 0 ms (beats 0%)
-// Memory   : 42176000 (beats 0%)
+// Memory   : 42604000 (beats 0%)
 // Language : java
 // Copyright: (c) 2026 srinivaseswar. All rights reserved.
 // Synced by: leetie
@@ -13,22 +13,22 @@
 class Solution {
     public int maxRotateFunction(int[] nums) {
         int n = nums.length;
-        int sum = 0;
-        int f0 = 0;
+        long sum = 0;
+        long f0 = 0;
 
         for (int i = 0; i < n; i++) {
             sum += nums[i];
-            f0 += i * nums[i];
+            f0 += (long) i * nums[i];
         }
 
-        int maxF = f0;
-        int currentF = f0;
+        long maxF = f0;
+        long currentF = f0;
 
         for (int k = 1; k < n; k++) {
-            currentF = currentF + sum - n * nums[n - k];
+            currentF = currentF + sum - (long) n * nums[n - k];
             maxF = Math.max(maxF, currentF);
         }
 
-        return maxF;
+        return (int) maxF;
     }
 }
